backend/requirements.txt
transformers>=5.5.0
tokenizers>=0.21.0
peft==0.13.0
torch==2.8.0
torchvision==0.23.0
whisperx==3.7.4
accelerate==1.12.0
numpy==2.0.2
pandas
pyyaml
fastapi
uvicorn
python-multipart
python-dotenv
sqlalchemy>=2.0
huggingface_hub
llama-cpp-python
nemo_toolkit[asr]
